backend/requirements.txt
boto3==1.34.19
openai==1.10.0
twilio==8.12.0
python-dotenv==1.0.0
aws-lambda-powertools==2.30.0
